+++ tests/job_test.py
@@ -22,21 +22,6 @@ import os
 _REPORT = "heightWeightReport"
 
 
-def test_nersc_project(fc):
-    from sirepo import nersc
-    from pykern.pkunit import pkeq, pkok, pkre, pkfail, pkexcept
-
-    _NO_SUCH_PROJECT = "NO_SUCH_PROJECT"
-
-    with pkexcept(nersc.invalid_project_msg(_NO_SUCH_PROJECT)):
-        nersc.assert_project(_NO_SUCH_PROJECT)
-
-    pkeq(
-        nersc.sbatch_account(nersc.VALID_TEST_ACCOUNT),
-        nersc.assert_project(nersc.VALID_TEST_ACCOUNT),
-    )
-
-
 def test_runCancel(fc):
     from pykern import pkunit
     from pykern.pkdebug import pkdc, pkdp, pkdlog
